+++ ipfabric/tools/__init__.py
@@ -1,5 +1,6 @@
 from .configuration import DeviceConfigs
+from .restore_intents import RestoreIntents
 from .sites import UpdateSiteNames
 from .vulnerabilities import Vulnerabilities
 
-__all__ = [DeviceConfigs, UpdateSiteNames, Vulnerabilities]
+__all__ = ["DeviceConfigs", "UpdateSiteNames", "Vulnerabilities", "RestoreIntents"]